+++ Editor/CrossUguiToggleEditor.cs
@@ -10,7 +10,8 @@ namespace HiddenAchievement.CrossguardUi
     {
         private SerializedProperty _crossGroupProperty;
         private SerializedProperty _idProperty;
-        private SerializedProperty _eventProperty;
+        private SerializedProperty _valChangeEventProperty;
+        private SerializedProperty _navSelectedEventProperty;
         
         protected override void OnEnable()
         {
@@ -18,7 +19,8 @@ namespace HiddenAchievement.CrossguardUi
             
             _crossGroupProperty = serializedObject.FindProperty("_crossGroup");
             _idProperty = serializedObject.FindProperty("_id");
-            _eventProperty = serializedObject.FindProperty("OnValueChangedForId");
+            _valChangeEventProperty = serializedObject.FindProperty("OnValueChangedForId");
+            _navSelectedEventProperty = serializedObject.FindProperty("OnNavSelected");
         }
 
         public override void OnInspectorGUI()
@@ -30,7 +32,8 @@ namespace HiddenAchievement.CrossguardUi
             EditorGUILayout.PropertyField(_crossGroupProperty);
             EditorGUILayout.PropertyField(_idProperty);
             EditorGUILayout.Space();
-            EditorGUILayout.PropertyField(_eventProperty);
+            EditorGUILayout.PropertyField(_valChangeEventProperty);
+            EditorGUILayout.PropertyField(_navSelectedEventProperty);
             serializedObject.ApplyModifiedProperties();
         }
     }